API documentation

Hermes functions

These functions publish and subscribe to Hermes MQTT-topics.

Main.SnipsHermesQnD.subscribe2Intents — Function

subscribe2Intents(intents, callback; moreTopics = nothing)

Subscribe to one or a list of intents and listen forever and run the callback if a matching intent is recieved.

Arguments:

intents: Abstract String or List of Abstract Strings to define intents to subscribe. The intents will be expanded to topics (i.e. "hermes/intent/SwitchOnLight")

callback: Function to be executed for a incoming message

moreTopics: keyword arg to provide additional topics to subscribe (complete names of of topics).

Details:

The callback function has the signature f(topic, intentMessage), where topic is a String and intentMessage a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-format) or a String, if the payload is not valid JSON. The callback function is spawned and the function is listening to the MQTT server while the callback is executed.

Main.SnipsHermesQnD.subscribe2Topics — Function

subscribe2Topics(topics, callback)

Subscribe to one or a list of topics and listen forever and run the callback if a matching intent is recieved.

Arguments:

topics: Abstract String or List of Abstract Strings to define topics to subscribe.

callback: Function to be executed for a incoming message

See subscribe2Intents() for details.

Main.SnipsHermesQnD.publishStartSessionAction — Function

publishStartSessionAction(text; siteId = CURRENT_SITE_ID,
     intentFilter = nothing, sendIntentNotRecognized = false,
     customData = nothing)

MQTT publish start session with init action

Arguments:

siteId: ID of the site in which the session is started

text: text to be said via TTS

intentFilter: Optional Array of String - a list of intent names to restrict the NLU resolution of the answer of this query.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

customData: data to be sent to the service.

Main.SnipsHermesQnD.publishStartSessionNotification — Function

publishStartSessionNotification(text; siteId = CURRENT_SITE_ID,
                                customData = nothing)

MQTT publish start session with init notification

Arguments:

siteId: siteID

text: text to be said via TTS

customData: data to be sent to the service.

Main.SnipsHermesQnD.publishEndSession — Function

publishEndSession(text; sessionId = CURRENT_SESSION_ID)

MQTT publish end session.

Arguments:

sessionId: ID of the session to be terminated as String. If omitted, sessionId of the current will be inserted.

text: text to be said via TTS

Dialogue manager functions

In addition to functions to work with the dialogue manager, advanced direct dialogues are provided that can be included in the control flow of the program.

Main.SnipsHermesQnD.publishContinueSession — Function

publishContinueSession(text; sessionId = CURRENT_SESSION_ID,
     intentFilter = nothing,
     customData = nothing, sendIntentNotRecognized = false)

MQTT publish continue session.

Arguments:

sessionId: ID of the current session as String

text: text to be said via TTS

intentFilter: Optional Array of String - a list of intents names to restrict the NLU resolution on the answer of this query.

customData: Optional String - an update to the session's custom data.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

Main.SnipsHermesQnD.listenIntentsOneTime — Function

listenIntentsOneTime(intents; moreTopics = nothing)

Subscribe to one or a list of Intents, but listen only until one matching intent is recognised.

Arguments

intents: AbstractString or List of AbstractString to define intents to subscribe or nothing

moreTopics: keyword arg to provide additional topics to subscribe to.

Value:

Return values are topic (as String) and payload (as Dict or as String if JSON parsing is not possible). If the topic is an intent, only the intent id is returned (i.e.: devname:intentname without the leading hermes/intent/)

Main.SnipsHermesQnD.configureIntent — Function

configureIntent(intent, on)

Enable or disable an intent.

Arguments:

intent: one intent to be configured

on: boolean value; if true, the intent is enabled; if false it is disabled.

Main.SnipsHermesQnD.askYesOrNoOrUnknown — Function

askYesOrNoOrUnknown(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :yes if "Yes" is answered or :no if "No" or :unknown otherwise.

Arguments:

question: String with the question to be uttered by Snips

Main.SnipsHermesQnD.askYesOrNo — Function

askYesOrNo(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :true if "Yes" or "No" otherwise.

Arguments:

question: String with the question to uttered

Main.SnipsHermesQnD.publishSay — Function

publishSay(text; sessionId = CURRENT_SESSION_ID, siteId = nothing,
                lang = LANG, id = nothing, wait = true)

Let the TTS say something.

The variant with a Symbol as first argument looks up the phrase in the dictionary of phrases for the selected language by calling getText().

Arguments:

text: text to be said via TTS

lang: optional language code to use when saying the text. If not specified, default language will be used

sessionId: optional ID of the session if there is one

id: optional request identifier. If provided, it will be passed back in the response on hermes/tts/sayFinished.

wait: wait until the massege is spoken (i.i. wait for the MQTT-topic)

Functions to handle intents

Main.SnipsHermesQnD.registerIntentAction — Function

registerIntentAction(intent, developer, inModule, action)
registerIntentAction(intent, action)

Add an intent to the list of intents to subscribe to. Each function that shall be executed if Snips recognises an intent must be registered with this function. The framework will collect all these links, subscribe to all needed intents and execute the respective functions. The links need not to be unique (in both directions): It is possible to assign several functions to one intent (all of them will be executed), or to assign one function to more then one intent.

The variant with only (intent, action) as arguments applies the variables CURRENTDEVELNAME and CURRENT_MODULE as stored in the framework. The variants registerIntent... create topics with prefix hermes/intent/developer:intent.

Arguments:

intent: Name of the intend (without developer name)

developer: Name of skill developer

inModule: current module (can be accessed with @__MODULE__)

action: the function to be linked with the intent

Main.SnipsHermesQnD.registerTriggerAction — Function

registerTriggerAction(intent, developer, inModule, action)
registerTriggerAction(intent, action)

Add an intent to the list of intents to subscribe to. Each function that shall be executed if Snips recognises The variants registerTrigger... create topics with prefix QnD/trigger/developer:intent.

See registerIntentAction() for details.

Main.SnipsHermesQnD.getIntentActions — Function

getIntentActions()

Return the list of all intent-function mappings for this app. The function is exported to deliver the mappings to the Main context.

Main.SnipsHermesQnD.setIntentActions — Function

setIntentActions(intentActions)

Overwrite the complete list of all intent-function mappings for this app. The function is exported to get the mappings from the Main context.

Arguments:

intentActions: Array of intent-action mappings as Tuple of (intent::AbstractString, developer::AbstractString, inModule::Module, action::Function)

Main.SnipsHermesQnD.publishSystemTrigger — Function

publishSystemTrigger(topic, trigger; develName = CURRENT_DEVEL_NAME)

Publish a system trigger with topic and payload.

Arguments:

topic: MQTT topic, with or w/o the developername. If no developername is included, CURRENTDEVELNAME will be added. If the topic does not start with qnd/trigger/, this will be added.

trigger: specific payload for the trigger.

Main.SnipsHermesQnD.makeSystemTrigger — Function

makeSystemTrigger(topic, trigger; develName = CURRENT_DEVEL_NAME)

Return (topic, payload) where topic is the fully quallified topic and payload a Dict() that include a system trigger topic and payload.

Arguments:

topic: MQTT topic, with or w/o the developername. If no developername is included, CURRENTDEVELNAME will be added. If the topic does not start with qnd/trigger/, this will be added.

trigger: specific payload for the trigger.

Main.SnipsHermesQnD.publishListenTrigger — Function

publishListenTrigger(mode)

Publish a stop-listen or start-listen trigger to make the assistant stop listening to voice commands.

Arguments:

mode: one of :stop or :start

Details:

The Skill AdoSnipsDoNotListen must be installed in order to respond to the trigger,otherwise the trigger will be ignored.

The trigger can be used to avoid false activation while watching TV or listening to the radio. Just publish the trigger as part of the "watch-TV-command". The trigger will disable all intents, listed in the config.ini and enable the listen-again intent only. The listen-again intent is double-checking any voice activation, so that only exact matches of commands (like "hör wieder zu" in German or "listen again" in English) will activate the intent.

The trigger must have the following JSON format: { "target" : "qnd/trigger/andreasdominik:ADoSnipsListen", "origin" : "ADoSnipsScheduler", "sessionId": "1234567890abcdef", "siteId" : "default", "time" : "timeString", "trigger" : { "command" : "stop" // or "start" } }

Main.SnipsHermesQnD.isFalseDetection — Function

isFalseDetection(payload)

Return true, if the current intent is not correct for the uttered command (i.e. false positive).

Arguments:

payload: Dictionary with the payload of a recognised intent.

Details:

All lines of the config.ini are analysed, witch match expressions:

`<intentname>:must_include:<description>=<list of words>

An example would be:

switchOnOff:must_include:just_words1=on,light

switchOnOff:must_chain:in_order=light,on

switchOnOff:must_span:start_to_end=switch,light,on

switchOnOff:must_span:start_to_end=switch,on

The command must include all words in the correct order of at least one parameter lines and the words must span the complete line (i.e. the command starts with first word and ends with the last word of the list).

Several lines are possible; the last part of the parameter name is used as description and to make the parameter names unique.

config.ini functions

Helper functions for read values from the file config.ini.

config.ini files follow the normal rules as for all Snips apps, with some extensions:

as for standard Snips, no spaces are allowed around the =

the parameter value may contain whitespace; i.e. light=my light returns one parameter with value: my light.

if the value of the parameter (right side) includes commas, the value can be interpreted as a comma-separated list of values. In this case, the reader-function will return an array of Strings with the values (which an be accessed by their index).

parameter names may have a prefix (set by the function setConfigPrefix()). If set, all config-functions will try to find parameter names with prefix. Example: If the config.ini includes the lines:

main_light:ip=192.168.0.15
wall_light:ip=192.168.0.16

the following code returns 192.168.0.15 in the first call and 192.168.0.16 in the second. This makes it easy to delegate config-reads to sub-functions:

Snips.setConfigPrefix("main_light")
main_ip = Snips.getConfig("ip")

Snips.setConfigPrefix("wall_light")
wall_ip = Snips.getConfig("ip")

Obviously is possible to access the parameters directly via Snips.setConfigPrefix("main_light:ip") or Snips.setConfigPrefix(Symbol("main_light:ip")) without setting the prefix (the second form will work even if another prefix is set; see doc of the functions for details).

Main.SnipsHermesQnD.getConfig — Function

getConfig(name; multiple = false, onePrefix = nothing)

Return the parameter value of the config.ini with name or nothing if the param does not exist. Return value is of type AbstractString, if it is a single value or of type AbstractArray{AbstractString} if the a list of values is read.

Arguments:

name: name of the config parameter as Symbol or String

multiple: if true an array of values is returned, even if only a single value have been read.

onePrefix: if defined, the prefix will be used only for this single call instead of the stored prefix.

Details:

If name is of type Symbol, it is treated as key in the Dirctionary of parameter values. If name is an AbstractString, the prefix is added if a prefix is defined (as <prefix>:<name>).

Main.SnipsHermesQnD.setConfigPrefix — Function

setConfigPrefix(newPrefix)

Set the prefix for all following calls to a parameter from config.ini. All parameter names, gives as Strings will be modified as <prefix>:<name>.

Main.SnipsHermesQnD.resetConfigPrefix — Function

resetConfigPrefix()

Remove the prefix for all following calls to a parameter from config.ini.

Main.SnipsHermesQnD.getAllConfig — Function

getAllConfig()

Return a Dict with the complete config.ini.

Main.SnipsHermesQnD.readConfig — Function

readConfig(appDir)

Read the lines of the App's config file and return a Dict with config values.

Arguments:

appDir: Directory of the currently running app.

Main.SnipsHermesQnD.matchConfig — Function

matchConfig(name::Symbol, val::String)

Return true if the parameter with name name of the config.ini has the value val or one element of the list as the value val.

Arguments:

name: name of the config parameter as Symbol or String

val: desired value

Main.SnipsHermesQnD.isInConfig — Function

isInConfig(name)

Return true if a parameter with name exists.

Arguments:

name: name of the config parameter as Symbol or String

Main.SnipsHermesQnD.isConfigValid — Function

isConfigValid(name; regex = r".", elem = 1, errorMsg = TEXTS[:error_config])

Return true, if the parameter name have been read correctly from the config.ini file and false otherwise. By default "correct" means: it is aString with length > 0. For a meire specific test, a regex can be provided.

Arguments:

name: name of parameter as AbstractString or Symbol

regex: optional regex for the test

elem: element to be tested, if the parameter returns an array

errorMsg: alternative error message.

Slot access functions

Functions to read values from slots of recognised intents.

Main.SnipsHermesQnD.extractSlotValue — Function

extractSlotValue(payload, slotName; multiple = false)

Return the value of a slot.

Nothing is returned, if

no slots in payload,

no slots with name slotName in payload,

no values in slot slotName.

If multiple == true, a list of all slot values will be returned. If false, only the 1st one as String.

Main.SnipsHermesQnD.isInSlot — Function

isInSlot(payload, slotName, value)

Return true, if the value is present in the slot slotName of the JSON payload (i.e. one of the slot values must match). Return false if something is wrong (value not in payload or no slots slotName.)

Main.SnipsHermesQnD.isOnOffMatched — Function

isOnOffMatched(payload, deviceName; siteId = CURRENT_SITE_ID)

Action to be combined with the ADoSnipsOnOFF intent. Depending on the payload the function returns:

:on if "on"

:off if "off"

:matched, if the device is matched but no on or off

:unmatched, if one of

wrong siteId

wrong device

Arguments:

payload: payload of intent

siteId: siteId of the device to be matched with the payload of intent if siteId == "any", the device will be matched w/o caring about siteId or room.

deviceName : name of device to be matched with the payload of intent

Main.SnipsHermesQnD.isValidOrEnd — Function

isValidOrEnd(param; errorMsg = "parameter is nothing")

End the session, with the message, if the param is nothing and returns false or true otherwise.

Function is a shortcut:

if param == nothing
    Snips.publishEndSession(:error)
    return true
end

is:

Snips.isValidOrEnd(param, :error) || return true

Arguments:

param: any value (from slot or config.ini) that may be nothing

errorMsg: Error as string or key of a text in the languages Dict (::Symbol).

Main.SnipsHermesQnD.readTimeFromSlot — Function

readTimeFromSlot(payload, slotName)

Return a DateTime from a slot with a time string of the format "2019-09-03 18:00:00 +00:00" or nothing if it is not possible to parse the slot.

MQTT functions

Low-level API to MQTT messages (publish and subscribe). In the QuickAndDirty framework, these functions are calling Eclipse mosquitto_pub and mosquitto_sub. However this first (and preliminary) implementation is surpriningly robust and easy to maintain - there might be no need to change.

Main.SnipsHermesQnD.subscribeMQTT — Function

subscribeMQTT(topics, callback; hostname = nothing, port = nothing)

Listen to one or more topics.

Arguments

topics: AbstractString or List of AbtsractString to define topics to subscribe

callback: Function to be executed for a incoming message.

hostname:

port: Hostname and port to listen. If not specified mosquitto_sub will be called without hostname/port (using the default configuration of the system).

Details:

The callback function has the signature f(topic, payload), where topic is a String and payload a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-format) or a String, if the payload is not a valid JSON.

The callback function is spawned.

Main.SnipsHermesQnD.readOneMQTT — Function

readOneMQTT(topics; hostname = nothing, port = nothing)

Listen to one or more topics until one message is retrieved and return topic as string and payload as Dict or as String if JSON parsing is not possible).

Arguments

topics: AbstractString or List of AbstractString to define topics to subscribe

hostname:

port: Hostname and port to listen. If not specified mosquitto_sub will be called without hostname/port (using the default configuration of the system).

Main.SnipsHermesQnD.publishMQTT — Function

publishMQTT(topic, payload, hostname = nothing, port = nothing)

Publish a MQTT message.

Arguments

topics: String with the topic

payload: Dict() with message

hostname:

port: Hostname and port to use. If not specified mosquitto_sub will be called without hostname/port (using the default configuration of the system).

Main.SnipsHermesQnD.publishMQTTfile — Function

publishMQTTfile(topic, fname, hostname = nothing, port = nothing)

Publish a MQTT message with a file as payload.

Arguments

topics: String with the topic

fname: full path and name of file to be published

hostname:

port: Hostname and port to use. If not specified mosquitto_sub will be called without hostname/port (using the default configuration of the system).

Handle background information of recognised intent

Main.SnipsHermesQnD.setSiteId — Function

setSiteId(siteId)

Set the siteId in the Module SnipsHermesQnD (necessary to direct the say() output to the current room)

Main.SnipsHermesQnD.getSiteId — Function

getSiteId()

Return the siteId in the Module SnipsHermesQnD (necessary to direct the say() output to the current room)

Main.SnipsHermesQnD.setSessionId — Function

setSessionId(sessionId)

Set the sessionId in the Module SnipsHermesQnD. The sessionId will be used to publish Hermes messages inside a runing session. The framework handles this in the background.

Arguments:

sessionId: as String from a Hermes payload.

Main.SnipsHermesQnD.getSessionId — Function

getSessionId()

Return the sessionId of the currently running session.

Main.SnipsHermesQnD.setDeveloperName — Function

setDeveloperName(name)

Set the developer name of the currently running app in the Module SnipsHermesQnD. The framework adds the name to MQTT messages in the background.

Arguments:

developerName: Name of the developer of the current app i.e. the part before the colon of an intent name.

Main.SnipsHermesQnD.getDeveloperName — Function

getDeveloperName()

Return the name of the develpper of the currently running app.

Main.SnipsHermesQnD.setModule — Function

setModule(currentModule)

Set the module of the currently running app in SnipsHermesQnD. The framework uses this in the background.

Arguments:

currentModule: The module in which the current skill is running. (acessible via marco @__MODULE__)

Main.SnipsHermesQnD.getModule — Function

getModule()

Return the module of the currently running app.

Main.SnipsHermesQnD.setAppDir — Function

setAppDir(appDir)

Store the directory appDir as CURRENTAPPDIR in the current session

Main.SnipsHermesQnD.getAppDir — Function

getAppDir()

Return the directory of the currently running app (i.e. the variable CURRENTAPPDIR)

Main.SnipsHermesQnD.setAppName — Function

setAppName(appName)

Store the name of the current app/module as CURRENTAPPNAME in the current session

Main.SnipsHermesQnD.getAppName — Function

getAppName()

Return the name of the currently running app (i.e. the variable CURRENTAPPNAME)

Main.SnipsHermesQnD.setTopic — Function

setTopic(topic)

Set the topic for which the currently running app is working. The framework uses this in the background.

Arguments:

topic: name of current topic

Main.SnipsHermesQnD.getTopic — Function

getTopic()

Return the topic of the currently running app.

Main.SnipsHermesQnD.setIntent — Function

setIntent(topic)

Set the intent for which the currently running app is working. The framework uses this in the background.

Arguments:

topic: name of current topic

Main.SnipsHermesQnD.getIntent — Function

getIntent()

Return the intent name of the currently running app.

Multi-language utilities

Main.SnipsHermesQnD.setLanguage — Function

setLanguage(lang)

Set the default language for SnipsHermesQnD. Currently supported laguages are "en" and "de".

This will affect publishSay() and all system messages. Log-messages will always be in English.

Arguments

lang: one of "en" or "de".

Main.SnipsHermesQnD.addText — Function

addText(lang::AbstractString, key::Symbol, text)

Add the text to the dictionary of text sniplets for the language lang and the key key.

Arguments:

lang: String with language code ("en", "de", ...)

key: Symbol with unique key for the text

text: String or array of String with the text(s) to be uttered.

Details:

If text is an Array, all texts will be saved and the function Snips.langText() will return a randomly selected text from the list.

If the key already exists, the new text will be added to the the Array of texts for a key.

Main.SnipsHermesQnD.langText — Function

langText(key::Symbol)
langText(key::Nothing)
langText(key::AbstractString)

Return the text in the languages dictionary for the key and the language set with setLanguage().

If the key does not exists, the text in the default language is returned; if this also does not exist, an error message is returned.

The variants make sure that nothing or the key itself are returned if key is nothing or an AbstractString, respectively.

Hardware control

Some devices can be directly controlled by the framework. In order to stay in the style of Snips, it is possible to run Shelly WiFi-switches without any cloud accounts and services. The Shelly-devices come with an own WiFi network. After installing the device just connect to Shelly's access point (somthing like shelly1-35FA58) and configure the switch for DHCP in your network with teh selft-explaining the web-interface of the device. At no point it's necessary to create an account or use a cloud service (although the Shelly1 documentation recommends).

Main.SnipsHermesQnD.setGPIO — Function

setGPIO(gpio, onoff::Symbol)

Switch a GPIO on or off with pigs.

Arguments:

gpio: ID of GPIO (not pinID)

onoff: one of :on or :off

Main.SnipsHermesQnD.switchShelly1 — Function

switchShelly1(ip, action; timer = 10)

Switch a shelly1 device with IP ip on or off, depending on the value given as action and return true, if successful.

Arguments:

ip: IP address or DNS name of Shelly1 device

action: demanded action as symbol; one of :on, :off, push or :timer. action :push will switch on for 200ms to simulate a push. action :timer will switch off after timer secs.

timer: if action == timer, the device is switched on and off again after timer secs (default 10s).

For the API-doc of the Shelly devices see: <https://shelly-api-docs.shelly.cloud>.

Main.SnipsHermesQnD.switchShelly25relay — Function

switchShelly25relay(ip, relay, action; timer = 10)

Switch the relay relay of a shelly2.5 device with IP ip on or off, depending on the value given as action and return true, if successful.

Arguments:

ip: IP address or DNS name of Shelly2.5 device

relay: Number of relay to be switched (0 or 1)

action: demanded action as symbol; one of :on, :off, push or :timer. action :push will switch on for 200ms to simulate a push. action :timer will switch off after timer secs.

timer: if action == timer, the device is switched on and off again after timer secs (default 10s).

For the API-doc of the Shelly devices see: <https://shelly-api-docs.shelly.cloud>.

Main.SnipsHermesQnD.moveShelly25roller — Function

moveShelly25roller(ip, action; pos = 100, duration = 1)

Move a roller with a shelly2.5 device with IP ip and return true, if successful.

Arguments:

ip: IP address or DNS name of Shelly2.5 device

action: demanded action as symbol; one of :open, :close, :stop or :to_pos.

pos: desired position in percent

'duration': is not yet implemented.

For the API-doc of the Shelly devices see: <https://shelly-api-docs.shelly.cloud>.

Status database

The framework handles a database to save status about house and devices, controlled by the assistant. The database is stored on disk in order to persist in case of a system crash or restart.

Every skill can store and read Dicts() as entries with a unique key or values as field-value-pairs as part of an entry.

The db looks somehow like:

{
    "irrigation" :
    {
        "time" : "2019-08-26T10:12:13.177"
        "writer" : "ADoSnipsIrrigation",
        "payload" :
        {
            "status" : "on",
            "next_status" : "off"
        }
    }
}

Location of the database file is <application_data_dir>/ADoSnipsQnD/<database_file> where application_data_dir and database_file> are parameters in the config.ini of the framework.

Main.SnipsHermesQnD.dbWritePayload — Function

dbWritePayload(key, payload)

Write a complete payload to a database entry. The payload is overwitten if the entry already exists, or created otherwise.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

payload: payload of the entry with key key to be written.

value: value to be stored in the field.

Main.SnipsHermesQnD.dbWriteValue — Function

dbWriteValue(key, field, value)

Write a field=>value pair to the payload of a database entry. The field is overwitten if the entry already exists, or created elsewise.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

field: database field of the payload of the entry with key key to be written (AbstractString or Symbol).

value: value to be stored in the field.

Main.SnipsHermesQnD.dbHasEntry — Function

dbHasEntry(key)

Check if the database has an entry with the key key and return true or false otherwise.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

Main.SnipsHermesQnD.dbReadEntry — Function

dbReadEntry(key)

Read the complete entry with the key key from the status database and return the entry as Dict() or nothing if not in the database.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

Main.SnipsHermesQnD.dbReadValue — Function

dbReadValue(key, field)

Read the field field of the entry with the key key from the status database and return the value or nothing if not in the database.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

field: database field of the payload of the entry with key key (AbstractString or Symbol).

Scheduler

The QnD framework provides a scheduler which allows to execute system triggers at a specified time in the future.

Schedules are added by sending a trigger with the following format to the scheduler. A list of triggers can be scheduled with one trigger:

{
  "origin": "ADoSnipsAuto",
  "topic": "qnd/trigger/andreasdominik:ADoSnipsSchedule",
  "siteId": "default",
  "sessionId": "7dab7a26-84fb-4855-8ad0-acd955408072",
  "trigger": {
    "mode": "add schedules",
    "sessionId": "7dab7a26-84fb-4855-8ad0-acd955408072",
    "siteId": "default",
    "time": "2019-08-26T14:07:55.623",
    "origin": "ADoSnipsAuto",
    "actions": [
      {
        "topic": "qnd/trigger/andreasdominik:ADoSnipsLights",
        "origin": "ADoSnipsAuto",
        "execute_time": "2019-08-28T10:00:20.534",
        "trigger": {
          "settings": "undefined",
          "device": "main_light",
          "onOrOff": "ON",
          "room": "default"
        }
      },
      {
        "topic": "qnd/trigger/andreasdominik:ADoSnipsLights",
        "origin": "ADoSnipsAuto",
        "execute_time": "2019-08-28T10:00:30.534",
        "trigger": {
          "settings": "undefined",
          "device": "main_light",
          "onOrOff": "OFF",
          "room": "default"
        }
      }
    ]
  }
}

A trigger with a mode od "delete all", "delete by topic" or "delete by origin" will delete the matching schedules:

{
  "origin": "Main.ADoSnipsTemplate",
  "topic": "qnd/trigger/andreasdominik:ADoSnipsSchedule",
  "siteId": "default",
  "sessionId": "7dab7a26-84fb-4855-8ad0-acd955408072",
  "trigger": {
    "mode": "delete all",
    "sessionId": "7dab7a26-84fb-4855-8ad0-acd955408072",
    "siteId": "default",
    "topic": "dummy",
    "origin": "dummy",
    "time": "2019-08-26T14:07:55.623"
  }
}

However, it is normally not necessary to set up these triggers manually; the following API functions provide an interface:

Main.SnipsHermesQnD.schedulerAddAction — Function

schedulerAddAction(executeTime, topic, trigger;
        sessionId = CURRENT_SESSION_ID,
        origin = CURRENT_APP_NAME,
        siteId = CURRENT_SITE_ID)

Add the trigger to the database of scheduled actions for execution at executeTime.

Arguments:

executeTime: DateTime object

topic: topic to which the system trigger will be published. topic has the format: "qnd/trigger/andreasdominik:ADoSnipsLights". The prefix "qnd/trigger/" and the developer name are added if missing in the argument.

trigger: The system trigger to be published as Dict(). Format of the trigger is defined by the target skill.

sessionId, origin and siteId defaults to the current values, if not given. SessionId and origin can be used to select scheduled actions for deletion.

Main.SnipsHermesQnD.schedulerAddActions — Function

schedulerAddActions(actions;
        sessionId = CURRENT_SESSION_ID,
        origin = CURRENT_APP_NAME,
        siteId = CURRENT_SITE_ID)

Add all actions in the list of action objects to the database of scheduled actions for execution. The elements of actions can be created by schedulerMakeAction() and must include executeTime, topic and the trigger to be published.

actions: List of actions to be published. Format of the trigger is defined by the target skill.

sessionId, origin and siteId defaults to the current values, if not given. SessionId and origin can be used to select scheduled actions for deletion.

Main.SnipsHermesQnD.schedulerMakeAction — Function

schedulerMakeAction(executeTime, topic, trigger;
                        sessionId = CURRENT_SESSION_ID,
                        origin = CURRENT_APP_NAME,
                        siteId = CURRENT_SITE_ID)

Return a Dict in the format for the QnD scheduler. A list of these object can be used to schedule many actions at once via schedulerAddActions(). (see documentation of schedulerAddAction() for details.)

Main.SnipsHermesQnD.schedulerDeleteAll — Function

schedulerDeleteAll()

Delete all scheduled action triggers.

Main.SnipsHermesQnD.schedulerDeleteTopic — Function

schedulerDeleteTopic(topic)

Delete all scheduled action triggers with the given topic.

Main.SnipsHermesQnD.schedulerDeleteOrigin — Function

schedulerDeleteOrigin(origin)

Delete all scheduled action triggers with the given origin (i.e. name of the app which cerated the scheduled action).

Utility functions

Little helpers to provide functionality which is commonly needed when developing a skill.

Main.SnipsHermesQnD.readableDateTime — Function

function readableDateTime(datetime::AbstractString; lang = LANG)
                          wholeDay = false, onlyDay = false)
function readableDateTime(datetime::DateTime; lang = LANG,
                          wholeDay = false, onlyDay = false)
function readableDate(datetime::Date; lang = LANG)

Return human readable date and time, like "Friday, 1. January 2018", "9 15"

Supported languages: "en", "de", "fr".

Arguments:

datetime: date and optional time as ISO string or as DateTime object.

lang: Language code (e.g. "en" or "de")

wholeDay: boolean, if true, no time is returned, instead "whole day"

onlyDay: just tell the day, not the time.

Value:

String value with readable date and time.

Main.SnipsHermesQnD.getOpenWeather — Function

getOpenWeather()

Return a Dict with weather information from openweather.org. The config.ini of the framework must include the lines

openweather_api_key=1234567abcdef
openweather_city_id=2950159

with a valid app-key (available from openweather.org for free) and the id of a city.

Value:

The return value has the elements:

:pressure: pressure in hPa

:temperature: temperature in degrees Celsius

:windspeed: wind speed in meter/sec

:winddir: wind direction in degrees

:clouds: cloudiness in percent

:rain1h

:rain3h: rain in mm in the last 1 or 3 hours

:sunrise

:sunset: local time of sunrise/sunset as DateTime object

Main.SnipsHermesQnD.tryrun — Function

tryrun(cmd; wait = true, errorMsg = TEXTS[:error_script], silent = flase)

Try to run an external command and returns true if successful or false if not.

Arguments:

cmd: command to be executed on the shell

wait: if true, wait until the command has finished

errorMsg: AbstractString or key to multi-language dict with the error message.

silent: if true, no error is published, if something went wrong.

Main.SnipsHermesQnD.ping — Function

ping(ip; c = 1, W = 1)

Return true, if a ping to the ip-address (or name) is successful.

Arguments:

c: number of pings to send (default: 1)

W: timeout (default 1 sec)

Main.SnipsHermesQnD.tryReadTextfile — Function

tryReadTextfile(fname, errorMsg = TEXTS[:error_read])

Try to read a text file from file system and return the text as String or an String of length 0, if something went wrong.

Main.SnipsHermesQnD.tryParseJSONfile — Function

tryParseJSONfile(fname; quiet = false)

Parse a JSON file and return a hierarchy of Dicts with the content.

keys are changed to Symbol

on error, an empty Dict() is returned

Arguments:

fname: filename

quiet: if false Snips utters an error message

Main.SnipsHermesQnD.tryParseJSON — Function

tryParseJSON(text)

parses a JSON and returns a hierarchy of Dicts{Symbol, Any} and Arrays with the content or a string (text), if text is not a valid JSON, the raw string is returned.

Main.SnipsHermesQnD.tryMkJSON — Function

tryMkJSON(payload)

Create a JSON representation of the input (nested Dict or Array) or return an empty string if not possible.

Main.SnipsHermesQnD.printLog — Function

printLog(s)

Print the message The current App-name is printed as prefix.

Main.SnipsHermesQnD.printDebug — Function

printDebug(s)

Print the message only, if debug-mode is on. Debug-modes include

none: no debugging

logging: only printDebug() will print

no_parallel: logging is on and skill actions will will not be spawned (as a result, the listener is off-line while a skill-action is running).

Current App-name is printed as prefix.

Index

Main.SnipsHermesQnD.addText

Main.SnipsHermesQnD.askYesOrNo

Main.SnipsHermesQnD.askYesOrNoOrUnknown

Main.SnipsHermesQnD.configureIntent

Main.SnipsHermesQnD.dbHasEntry

Main.SnipsHermesQnD.dbReadEntry

Main.SnipsHermesQnD.dbReadValue

Main.SnipsHermesQnD.dbWritePayload

Main.SnipsHermesQnD.dbWriteValue

Main.SnipsHermesQnD.extractSlotValue

Main.SnipsHermesQnD.getAllConfig

Main.SnipsHermesQnD.getAppDir

Main.SnipsHermesQnD.getAppName

Main.SnipsHermesQnD.getConfig

Main.SnipsHermesQnD.getDeveloperName

Main.SnipsHermesQnD.getIntent

Main.SnipsHermesQnD.getIntentActions

Main.SnipsHermesQnD.getModule

Main.SnipsHermesQnD.getOpenWeather

Main.SnipsHermesQnD.getSessionId

Main.SnipsHermesQnD.getSiteId

Main.SnipsHermesQnD.getTopic

Main.SnipsHermesQnD.isConfigValid

Main.SnipsHermesQnD.isFalseDetection

Main.SnipsHermesQnD.isInConfig

Main.SnipsHermesQnD.isInSlot

Main.SnipsHermesQnD.isOnOffMatched

Main.SnipsHermesQnD.isValidOrEnd

Main.SnipsHermesQnD.langText

Main.SnipsHermesQnD.listenIntentsOneTime

Main.SnipsHermesQnD.makeSystemTrigger

Main.SnipsHermesQnD.matchConfig

Main.SnipsHermesQnD.moveShelly25roller

Main.SnipsHermesQnD.ping

Main.SnipsHermesQnD.printDebug

Main.SnipsHermesQnD.printLog

Main.SnipsHermesQnD.publishContinueSession

Main.SnipsHermesQnD.publishEndSession

Main.SnipsHermesQnD.publishListenTrigger

Main.SnipsHermesQnD.publishMQTT

Main.SnipsHermesQnD.publishMQTTfile

Main.SnipsHermesQnD.publishSay

Main.SnipsHermesQnD.publishStartSessionAction

Main.SnipsHermesQnD.publishStartSessionNotification

Main.SnipsHermesQnD.publishSystemTrigger

Main.SnipsHermesQnD.readConfig

Main.SnipsHermesQnD.readOneMQTT

Main.SnipsHermesQnD.readTimeFromSlot

Main.SnipsHermesQnD.readableDateTime

Main.SnipsHermesQnD.registerIntentAction

Main.SnipsHermesQnD.registerTriggerAction

Main.SnipsHermesQnD.resetConfigPrefix

Main.SnipsHermesQnD.schedulerAddAction

Main.SnipsHermesQnD.schedulerAddActions

Main.SnipsHermesQnD.schedulerDeleteAll

Main.SnipsHermesQnD.schedulerDeleteOrigin

Main.SnipsHermesQnD.schedulerDeleteTopic

Main.SnipsHermesQnD.schedulerMakeAction

Main.SnipsHermesQnD.setAppDir

Main.SnipsHermesQnD.setAppName

Main.SnipsHermesQnD.setConfigPrefix

Main.SnipsHermesQnD.setDeveloperName

Main.SnipsHermesQnD.setGPIO

Main.SnipsHermesQnD.setIntent

Main.SnipsHermesQnD.setIntentActions

Main.SnipsHermesQnD.setLanguage

Main.SnipsHermesQnD.setModule

Main.SnipsHermesQnD.setSessionId

Main.SnipsHermesQnD.setSiteId

Main.SnipsHermesQnD.setTopic

Main.SnipsHermesQnD.subscribe2Intents

Main.SnipsHermesQnD.subscribe2Topics

Main.SnipsHermesQnD.subscribeMQTT

Main.SnipsHermesQnD.switchShelly1

Main.SnipsHermesQnD.switchShelly25relay

Main.SnipsHermesQnD.tryMkJSON

Main.SnipsHermesQnD.tryParseJSON

Main.SnipsHermesQnD.tryParseJSONfile

Main.SnipsHermesQnD.tryReadTextfile

Main.SnipsHermesQnD.tryrun